fix architecture
