+++ src/bootstrap/__init__.py
@@ -1,0 +1,1 @@
+"""Application composition root helpers."""
